greatly reduced complexity of navigatioon
