+++ run_tests
@@ -5,7 +5,7 @@ echo "------------------------------------ Utilites ----------------------------
 echo ""
 rosrun tfr_utilities tfr_utilities-test
 echo ""
-echo "----------------------------------- Navigation -----------------------------------"
 echo ""
-rosrun tfr_navigation tfr_navigation-test
-
+echo "------------------------------------- Control -------------------------------------"
+echo ""
+rosrun tfr_control tfr_control-test
